+++ tencentcloud/ckafka/v20190819/models.go
@@ -1763,6 +1763,24 @@ func (r *CreateDatahubTopicResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }
 
+type CreateInstancePostData struct {
+	// CreateInstancePre返回固定为0，不能作为CheckTaskStatus的查询条件。只是为了保证和后台数据结构对齐。
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	FlowId *int64 `json:"FlowId,omitempty" name:"FlowId"`
+
+	// 订单号列表
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	DealNames []*string `json:"DealNames,omitempty" name:"DealNames"`
+
+	// 实例Id，当购买多个实例时，默认返回购买的第一个实例 id
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	InstanceId *string `json:"InstanceId,omitempty" name:"InstanceId"`
+
+	// 订单和购买实例对应映射列表
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	DealNameInstanceIdMapping []*DealInstanceDTO `json:"DealNameInstanceIdMapping,omitempty" name:"DealNameInstanceIdMapping"`
+}
+
 // Predefined struct for user
 type CreateInstancePostRequestParams struct {
 	// 实例名称，是一个不超过 64 个字符的字符串，必须以字母为首字符，剩余部分可以包含字母、数字和横划线(-)
@@ -1912,6 +1930,18 @@ func (r *CreateInstancePostRequest) FromJsonString(s string) error {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreateInstancePostRequest has unknown keys!", "")
 	}
 	return json.Unmarshal([]byte(s), &r)
+}
+
+type CreateInstancePostResp struct {
+	// 返回的code，0为正常，非0为错误
+	ReturnCode *string `json:"ReturnCode,omitempty" name:"ReturnCode"`
+
+	// 接口返回消息，当接口报错时提示错误信息
+	ReturnMessage *string `json:"ReturnMessage,omitempty" name:"ReturnMessage"`
+
+	// 返回的Data数据
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	Data *CreateInstancePostData `json:"Data,omitempty" name:"Data"`
 }
 
 // Predefined struct for user
@@ -2228,6 +2258,182 @@ func (r *CreatePartitionResponse) ToJsonString() string {
 // FromJsonString It is highly **NOT** recommended to use this function
 // because it has no param check, nor strict type check
 func (r *CreatePartitionResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type CreatePostPaidInstanceRequestParams struct {
+	// 实例名称，是一个不超过 64 个字符的字符串，必须以字母为首字符，剩余部分可以包含字母、数字和横划线(-)
+	InstanceName *string `json:"InstanceName,omitempty" name:"InstanceName"`
+
+	// 创建的实例默认接入点所在的 vpc 对应 vpcId。目前不支持创建基础网络实例，因此该参数必填
+	VpcId *string `json:"VpcId,omitempty" name:"VpcId"`
+
+	// 子网id。创建实例默认接入点所在的子网对应的子网 id
+	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
+
+	// 实例规格。当创建标准版实例时必填，创建专业版实例时不需要填写。1：入门型；2：标准型；3：进阶型；4：容量型；5：高阶型1；6：高阶性2；7：高阶型3；8：高阶型4；9 ：独占型
+	InstanceType *int64 `json:"InstanceType,omitempty" name:"InstanceType"`
+
+	// 实例日志的默认最长保留时间，单位分钟。不传入该参数时默认为 1440 分钟（1天），最大30天。当 topic 显式设置消息保留时间时，以 topic 保留时间为准
+	MsgRetentionTime *int64 `json:"MsgRetentionTime,omitempty" name:"MsgRetentionTime"`
+
+	// 创建实例时可以选择集群Id, 该入参表示集群Id。不指定实例所在集群则不传入该参数
+	ClusterId *int64 `json:"ClusterId,omitempty" name:"ClusterId"`
+
+	// 实例版本。目前支持 "0.10.2","1.1.1","2.4.2","2.8.1"
+	KafkaVersion *string `json:"KafkaVersion,omitempty" name:"KafkaVersion"`
+
+	// 实例类型。"standard"：标准版，"profession"：专业版
+	SpecificationsType *string `json:"SpecificationsType,omitempty" name:"SpecificationsType"`
+
+	// 实例硬盘类型，"CLOUD_BASIC"：云硬盘，"CLOUD_SSD"：高速云硬盘。不传默认为 "CLOUD_BASIC"
+	DiskType *string `json:"DiskType,omitempty" name:"DiskType"`
+
+	// 实例内网峰值带宽。单位 MB/s。标准版需传入当前实例规格所对应的峰值带宽。注意如果创建的实例为专业版实例，峰值带宽，分区数等参数配置需要满足专业版的计费规格。
+	BandWidth *int64 `json:"BandWidth,omitempty" name:"BandWidth"`
+
+	// 实例硬盘大小，需要满足当前实例的计费规格
+	DiskSize *int64 `json:"DiskSize,omitempty" name:"DiskSize"`
+
+	// 实例最大分区数量，需要满足当前实例的计费规格
+	Partition *int64 `json:"Partition,omitempty" name:"Partition"`
+
+	// 实例最大 topic 数量，需要满足当前实例的计费规格
+	TopicNum *int64 `json:"TopicNum,omitempty" name:"TopicNum"`
+
+	// 实例所在的可用区。当创建多可用区实例时，该参数为创建的默认接入点所在子网的可用区 id
+	ZoneId *int64 `json:"ZoneId,omitempty" name:"ZoneId"`
+
+	// 当前实例是否为多可用区实例。
+	MultiZoneFlag *bool `json:"MultiZoneFlag,omitempty" name:"MultiZoneFlag"`
+
+	// 当实例为多可用区实例时，多可用区 id 列表。注意参数 ZoneId 对应的多可用区需要包含在该参数数组中
+	ZoneIds []*int64 `json:"ZoneIds,omitempty" name:"ZoneIds"`
+
+	// 购买实例数量。非必填，默认值为 1。当传入该参数时，会创建多个 instanceName 加后缀区分的实例
+	InstanceNum *int64 `json:"InstanceNum,omitempty" name:"InstanceNum"`
+
+	// 公网带宽大小，单位 Mbps。默认是没有加上免费 3Mbps 带宽。例如总共需要 3Mbps 公网带宽，此处传 0；总共需要 4Mbps 公网带宽，此处传 1
+	PublicNetworkMonthly *int64 `json:"PublicNetworkMonthly,omitempty" name:"PublicNetworkMonthly"`
+}
+
+type CreatePostPaidInstanceRequest struct {
+	*tchttp.BaseRequest
+	
+	// 实例名称，是一个不超过 64 个字符的字符串，必须以字母为首字符，剩余部分可以包含字母、数字和横划线(-)
+	InstanceName *string `json:"InstanceName,omitempty" name:"InstanceName"`
+
+	// 创建的实例默认接入点所在的 vpc 对应 vpcId。目前不支持创建基础网络实例，因此该参数必填
+	VpcId *string `json:"VpcId,omitempty" name:"VpcId"`
+
+	// 子网id。创建实例默认接入点所在的子网对应的子网 id
+	SubnetId *string `json:"SubnetId,omitempty" name:"SubnetId"`
+
+	// 实例规格。当创建标准版实例时必填，创建专业版实例时不需要填写。1：入门型；2：标准型；3：进阶型；4：容量型；5：高阶型1；6：高阶性2；7：高阶型3；8：高阶型4；9 ：独占型
+	InstanceType *int64 `json:"InstanceType,omitempty" name:"InstanceType"`
+
+	// 实例日志的默认最长保留时间，单位分钟。不传入该参数时默认为 1440 分钟（1天），最大30天。当 topic 显式设置消息保留时间时，以 topic 保留时间为准
+	MsgRetentionTime *int64 `json:"MsgRetentionTime,omitempty" name:"MsgRetentionTime"`
+
+	// 创建实例时可以选择集群Id, 该入参表示集群Id。不指定实例所在集群则不传入该参数
+	ClusterId *int64 `json:"ClusterId,omitempty" name:"ClusterId"`
+
+	// 实例版本。目前支持 "0.10.2","1.1.1","2.4.2","2.8.1"
+	KafkaVersion *string `json:"KafkaVersion,omitempty" name:"KafkaVersion"`
+
+	// 实例类型。"standard"：标准版，"profession"：专业版
+	SpecificationsType *string `json:"SpecificationsType,omitempty" name:"SpecificationsType"`
+
+	// 实例硬盘类型，"CLOUD_BASIC"：云硬盘，"CLOUD_SSD"：高速云硬盘。不传默认为 "CLOUD_BASIC"
+	DiskType *string `json:"DiskType,omitempty" name:"DiskType"`
+
+	// 实例内网峰值带宽。单位 MB/s。标准版需传入当前实例规格所对应的峰值带宽。注意如果创建的实例为专业版实例，峰值带宽，分区数等参数配置需要满足专业版的计费规格。
+	BandWidth *int64 `json:"BandWidth,omitempty" name:"BandWidth"`
+
+	// 实例硬盘大小，需要满足当前实例的计费规格
+	DiskSize *int64 `json:"DiskSize,omitempty" name:"DiskSize"`
+
+	// 实例最大分区数量，需要满足当前实例的计费规格
+	Partition *int64 `json:"Partition,omitempty" name:"Partition"`
+
+	// 实例最大 topic 数量，需要满足当前实例的计费规格
+	TopicNum *int64 `json:"TopicNum,omitempty" name:"TopicNum"`
+
+	// 实例所在的可用区。当创建多可用区实例时，该参数为创建的默认接入点所在子网的可用区 id
+	ZoneId *int64 `json:"ZoneId,omitempty" name:"ZoneId"`
+
+	// 当前实例是否为多可用区实例。
+	MultiZoneFlag *bool `json:"MultiZoneFlag,omitempty" name:"MultiZoneFlag"`
+
+	// 当实例为多可用区实例时，多可用区 id 列表。注意参数 ZoneId 对应的多可用区需要包含在该参数数组中
+	ZoneIds []*int64 `json:"ZoneIds,omitempty" name:"ZoneIds"`
+
+	// 购买实例数量。非必填，默认值为 1。当传入该参数时，会创建多个 instanceName 加后缀区分的实例
+	InstanceNum *int64 `json:"InstanceNum,omitempty" name:"InstanceNum"`
+
+	// 公网带宽大小，单位 Mbps。默认是没有加上免费 3Mbps 带宽。例如总共需要 3Mbps 公网带宽，此处传 0；总共需要 4Mbps 公网带宽，此处传 1
+	PublicNetworkMonthly *int64 `json:"PublicNetworkMonthly,omitempty" name:"PublicNetworkMonthly"`
+}
+
+func (r *CreatePostPaidInstanceRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *CreatePostPaidInstanceRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "InstanceName")
+	delete(f, "VpcId")
+	delete(f, "SubnetId")
+	delete(f, "InstanceType")
+	delete(f, "MsgRetentionTime")
+	delete(f, "ClusterId")
+	delete(f, "KafkaVersion")
+	delete(f, "SpecificationsType")
+	delete(f, "DiskType")
+	delete(f, "BandWidth")
+	delete(f, "DiskSize")
+	delete(f, "Partition")
+	delete(f, "TopicNum")
+	delete(f, "ZoneId")
+	delete(f, "MultiZoneFlag")
+	delete(f, "ZoneIds")
+	delete(f, "InstanceNum")
+	delete(f, "PublicNetworkMonthly")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "CreatePostPaidInstanceRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type CreatePostPaidInstanceResponseParams struct {
+	// 返回结果
+	Result *CreateInstancePostResp `json:"Result,omitempty" name:"Result"`
+
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type CreatePostPaidInstanceResponse struct {
+	*tchttp.BaseResponse
+	Response *CreatePostPaidInstanceResponseParams `json:"Response"`
+}
+
+func (r *CreatePostPaidInstanceResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *CreatePostPaidInstanceResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }