+++ tencentcloud/essbasic/v20210526/models.go
@@ -2761,6 +2761,13 @@ type Component struct {
 
 	// 扩展参数：
 	// 为JSON格式。
+	// 不同类型的控件会有部分非通用参数
+	// 
+	// TEXT/MULTI_LINE_TEXT控件可以指定
+	// 1 Font：目前只支持黑体、宋体
+	// 2 FontSize： 范围12-72
+	// 3 FontAlign： Left/Right/Center，左对齐/居中/右对齐
+	// 例如：{"FontSize":12}
 	// 
 	// ComponentType为FILL_IMAGE时，支持以下参数：
 	// NotMakeImageCenter：bool。是否设置图片居中。false：居中（默认）。 true: 不居中
@@ -2783,50 +2790,11 @@ type Component struct {
 	// 5 Gaps:： 字符串类型，仅在Format为“yyyy m d”时起作用，格式为用逗号分开的两个整数，例如”2,2”，两个数字分别是日期格式的前后两个空隙中的空格个数
 	// 如果extra参数为空，默认为”yyyy年m月d日”格式的居中日期
 	// 特别地，如果extra中Format字段为空或无法被识别，则extra参数会被当作默认值处理（Font，FontSize，Gaps和FontAlign都不会起效）
-	// 参数样例：    "ComponentExtra": "{\"Format\":“yyyy m d”,\"FontSize\":12,\"Gaps\":\"2,2\", \"FontAlign\":\"Right\"}"
+	// 参数样例： "ComponentExtra": "{"Format":“yyyy m d”,"FontSize":12,"Gaps":"2,2", "FontAlign":"Right"}"
 	// 
 	// ComponentType为SIGN_SEAL类型时，支持以下参数：
 	// 1.PageRanges：PageRange的数组，通过PageRanges属性设置该印章在PDF所有页面上盖章（适用于标书在所有页面盖章的情况）
-	// 参数样例： "ComponentExtra":"{\"PageRanges\":[\"PageRange\":{\"BeginPage\":1,\"EndPage\":-1}]}"
-	// 
-	// 
-	// 参数控件样式，json格式表述
-	// 
-	// 不同类型的控件会有部分非通用参数
-	// 
-	// TEXT/MULTI_LINE_TEXT控件可以指定
-	// 1 Font：目前只支持黑体、宋体
-	// 2 FontSize： 范围12-72
-	// 3 FontAlign： Left/Right/Center，左对齐/居中/右对齐
-	// 例如：{"FontSize":12}
-	// 
-	// ComponentType为FILL_IMAGE时，支持以下参数：
-	// NotMakeImageCenter：bool。是否设置图片居中。false：居中（默认）。 true: 不居中
-	// FillMethod: int. 填充方式。0-铺满（默认）；1-等比例缩放
-	// 
-	// ComponentType为SIGN_SIGNATURE类型可以控制签署方式
-	// {“ComponentTypeLimit”: [“xxx”]}
-	// xxx可以为：
-	// HANDWRITE – 手写签名
-	// BORDERLESS_ESIGN – 自动生成无边框腾讯体
-	// OCR_ESIGN -- AI智能识别手写签名
-	// ESIGN -- 个人印章类型
-	// SYSTEM_ESIGN -- 系统签名（该类型可以在用户签署时根据用户姓名一键生成一个签名来进行签署）
-	// 如：{“ComponentTypeLimit”: [“BORDERLESS_ESIGN”]}
-	// 
-	// ComponentType为SIGN_DATE时，支持以下参数：
-	// 1 Font：字符串类型目前只支持"黑体"、"宋体"，如果不填默认为"黑体"
-	// 2 FontSize： 数字类型，范围6-72，默认值为12
-	// 3 FontAlign： 字符串类型，可取Left/Right/Center，对应左对齐/居中/右对齐
-	// 4 Format： 字符串类型，日期格式，必须是以下五种之一 “yyyy m d”，”yyyy年m月d日”，”yyyy/m/d”，”yyyy-m-d”，”yyyy.m.d”。
-	// 5 Gaps:： 字符串类型，仅在Format为“yyyy m d”时起作用，格式为用逗号分开的两个整数，例如”2,2”，两个数字分别是日期格式的前后两个空隙钟的空格个数
-	// 如果extra参数为空，默认为”yyyy年m月d日”格式的居中日期
-	// 特别地，如果extra中Format字段为空或无法被识别，则extra参数会被当作默认值处理（Font，FontSize，Gaps和FontAlign都不会起效）
-	// 参数样例：    "ComponentExtra": "{\"Format\":“yyyy m d”,\"FontSize\":12,\"Gaps\":\"2,2\", \"FontAlign\":\"Right\"}"
-	// 
-	// ComponentType为SIGN_SEAL类型时，支持以下参数：
-	// 1.PageRanges：PageRange的数组，通过PageRanges属性设置该印章在PDF所有页面上盖章（适用于标书在所有页面盖章的情况）
-	// 参数样例： "ComponentExtra":"{\"PageRanges\":[\"PageRange\":{\"BeginPage\":1,\"EndPage\":-1}]}"
+	// 参数样例： "ComponentExtra":"{["PageRange":{"BeginPage":1,"EndPage":-1}]}"
 	ComponentExtra *string `json:"ComponentExtra,omitempty" name:"ComponentExtra"`
 
 	// 控件填充vaule，ComponentType和传入值类型对应关系：