+++ tencentcloud/postgres/v20170312/models.go
@@ -1282,7 +1282,7 @@ func (r *CreateParameterTemplateResponse) FromJsonString(s string) error {
 
 // Predefined struct for user
 type CreateReadOnlyDBInstanceRequestParams struct {
-	// 售卖规格ID。该参数可以通过调用DescribeProductConfig的返回值中的SpecCode字段来获取。
+	// 售卖规格ID。该参数可以通过调用DescribeClasses的返回值中的SpecCode字段来获取。
 	SpecCode *string `json:"SpecCode,omitempty" name:"SpecCode"`
 
 	// 实例容量大小，单位：GB。
@@ -1306,7 +1306,7 @@ type CreateReadOnlyDBInstanceRequestParams struct {
 	// 【废弃】不再需要指定，内核版本号与主实例保持一致
 	DBVersion *string `json:"DBVersion,omitempty" name:"DBVersion"`
 
-	// 实例计费类型。目前支持：PREPAID（预付费，即包年包月），POSTPAID_BY_HOUR（后付费，即按量计费）。如果主实例为后付费，只读实例必须也为后付费。
+	// 实例计费类型。目前支持：PREPAID（预付费，即包年包月），POSTPAID_BY_HOUR（后付费，即按量计费）。默认值：PREPAID。如果主实例为后付费，只读实例必须也为后付费。
 	InstanceChargeType *string `json:"InstanceChargeType,omitempty" name:"InstanceChargeType"`
 
 	// 是否自动使用代金券。1（是），0（否），默认不使用。
@@ -1346,7 +1346,7 @@ type CreateReadOnlyDBInstanceRequestParams struct {
 type CreateReadOnlyDBInstanceRequest struct {
 	*tchttp.BaseRequest
 	
-	// 售卖规格ID。该参数可以通过调用DescribeProductConfig的返回值中的SpecCode字段来获取。
+	// 售卖规格ID。该参数可以通过调用DescribeClasses的返回值中的SpecCode字段来获取。
 	SpecCode *string `json:"SpecCode,omitempty" name:"SpecCode"`
 
 	// 实例容量大小，单位：GB。
@@ -1370,7 +1370,7 @@ type CreateReadOnlyDBInstanceRequest struct {
 	// 【废弃】不再需要指定，内核版本号与主实例保持一致
 	DBVersion *string `json:"DBVersion,omitempty" name:"DBVersion"`
 
-	// 实例计费类型。目前支持：PREPAID（预付费，即包年包月），POSTPAID_BY_HOUR（后付费，即按量计费）。如果主实例为后付费，只读实例必须也为后付费。
+	// 实例计费类型。目前支持：PREPAID（预付费，即包年包月），POSTPAID_BY_HOUR（后付费，即按量计费）。默认值：PREPAID。如果主实例为后付费，只读实例必须也为后付费。
 	InstanceChargeType *string `json:"InstanceChargeType,omitempty" name:"InstanceChargeType"`
 
 	// 是否自动使用代金券。1（是），0（否），默认不使用。