+++ CHANGELOG.md
@@ -1,3 +1,257 @@
+# Release v1.0.687
+
+## 内容分发网络 CDN(cdn) 版本：2018-06-06
+
+### 第 124 次发布
+
+发布时间：2023-06-26 01:13:28
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [ModifyDomainConfig](https://cloud.tencent.com/document/api/228/94258)
+
+
+
+## 消息队列 CKafka(ckafka) 版本：2019-08-19
+
+### 第 85 次发布
+
+发布时间：2023-06-26 01:15:49
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [CreatePostPaidInstance](https://cloud.tencent.com/document/api/597/94259)
+
+新增数据结构：
+
+* [CreateInstancePostData](https://cloud.tencent.com/document/api/597/40861#CreateInstancePostData)
+* [CreateInstancePostResp](https://cloud.tencent.com/document/api/597/40861#CreateInstancePostResp)
+
+
+
+## TDSQL-C MySQL 版(cynosdb) 版本：2019-01-07
+
+### 第 58 次发布
+
+发布时间：2023-06-26 01:22:36
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [DescribeSupportProxyVersion](https://cloud.tencent.com/document/api/1003/94260)
+
+
+
+## 云数据库 SQL Server(sqlserver) 版本：2018-03-28
+
+### 第 46 次发布
+
+发布时间：2023-06-26 01:49:10
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DescribeDBInstancesAttribute](https://cloud.tencent.com/document/api/238/90299)
+
+	* 新增出参：TDEConfig
+
+* [DescribeDBs](https://cloud.tencent.com/document/api/238/19968)
+
+	* 新增入参：Encryption
+
+
+新增数据结构：
+
+* [TDEConfigAttribute](https://cloud.tencent.com/document/api/238/19976#TDEConfigAttribute)
+
+修改数据结构：
+
+* [DBDetail](https://cloud.tencent.com/document/api/238/19976#DBDetail)
+
+	* 新增成员：Encryption
+
+
+
+
+## 腾讯云区块链服务平台 TBaaS(tbaas) 版本：2018-04-16
+
+### 第 16 次发布
+
+发布时间：2023-06-26 01:51:21
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+<font color="#dd0000">**删除接口**：</font>
+
+* DeployDynamicBcosContract
+* GetBcosBlockByNumber
+* GetBcosBlockList
+* GetBcosTransByHash
+* GetBcosTransList
+* InvokeBcosTrans
+
+<font color="#dd0000">**删除数据结构**：</font>
+
+* BcosBlockObj
+* BcosTransInfo
+
+
+
+## 分布式身份(tdid) 版本：2021-05-19
+
+### 第 7 次发布
+
+发布时间：2023-06-26 01:54:56
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+<font color="#dd0000">**删除接口**：</font>
+
+* CancelAuthorityIssuer
+* CheckDidDeploy
+* CreateDidService
+* CreateLabel
+* DeployByName
+* DownCpt
+* EnableHash
+* GetAuthoritiesList
+* GetCptList
+* GetCredentialIssueRank
+* GetCredentialIssueTrend
+* GetDataPanel
+* GetDeployInfo
+* GetDeployList
+* GetDidDetail
+* GetDidList
+* GetDidRegisterTrend
+* GetDidServiceDetail
+* GetDidServiceList
+* GetGroupList
+* GetLabelList
+* GetPolicyList
+* GetPublicKey
+* QueryPolicy
+* RecognizeAuthorityIssuer
+* RegisterClaimPolicy
+* RegisterIssuer
+* RemoveHash
+
+<font color="#dd0000">**删除数据结构**：</font>
+
+* Authority
+* Contract
+* CptListData
+* DidData
+* DidServiceInfo
+* Group
+* Label
+* Policy
+* Task
+* Trend
+
+
+
+## TI-ONE 训练平台(tione) 版本：2021-11-11
+
+### 第 27 次发布
+
+发布时间：2023-06-26 01:57:45
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [CFSConfig](https://cloud.tencent.com/document/api/851/75051#CFSConfig)
+
+	* 新增成员：MountType, Protocol
+
+
+
+
+## TI-ONE 训练平台(tione) 版本：2019-10-22
+
+
+
+## T-Sec-安心平台(RP)(trp) 版本：2021-05-15
+
+### 第 22 次发布
+
+发布时间：2023-06-26 02:00:13
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateCodeBatch](https://cloud.tencent.com/document/api/1458/75029)
+
+	* 新增入参：ValidDate, ProductionDate
+
+* [DescribeScanLogs](https://cloud.tencent.com/document/api/1458/85943)
+
+	* 新增入参：PageSize, PageNumber, Openid
+
+	* <font color="#dd0000">**修改入参**：</font>Code
+
+* [DescribeTraceCodeById](https://cloud.tencent.com/document/api/1458/75010)
+
+	* 新增出参：CodePath
+
+* [ModifyCodeBatch](https://cloud.tencent.com/document/api/1458/75007)
+
+	* 新增入参：ValidDate, ProductionDate
+
+
+新增数据结构：
+
+* [AttrItem](https://cloud.tencent.com/document/api/1458/75030#AttrItem)
+
+修改数据结构：
+
+* [CodeBatch](https://cloud.tencent.com/document/api/1458/75030#CodeBatch)
+
+	* 新增成员：ProductionDate, ValidDate, Attrs
+
+* [Product](https://cloud.tencent.com/document/api/1458/75030#Product)
+
+	* <font color="#dd0000">**修改成员**：</font>ProductId, CorpId, ProductCode, Specification, Remark, Logo, CreateTime, UpdateTime, Ext, MerchantName
+
+* [ScanLog](https://cloud.tencent.com/document/api/1458/75030#ScanLog)
+
+	* 新增成员：Type, MerchantName, ProductName
+
+* [TraceData](https://cloud.tencent.com/document/api/1458/75030#TraceData)
+
+	* <font color="#dd0000">**修改成员**：</font>TraceItems
+
+* [TraceItem](https://cloud.tencent.com/document/api/1458/75030#TraceItem)
+
+	* <font color="#dd0000">**修改成员**：</font>ReadOnly, Hidden, Key, Ext
+
+
+
+
 # Release v1.0.686
 
 ## 云数据库 MySQL(cdb) 版本：2017-03-20